+++ spring-data-mongodb/src/main/java/org/springframework/data/mongodb/aot/MongoRuntimeHints.java
@@ -31,8 +31,6 @@ import org.springframework.data.mongodb.core.mapping.event.ReactiveAfterConvertC
 import org.springframework.data.mongodb.core.mapping.event.ReactiveAfterSaveCallback;
 import org.springframework.data.mongodb.core.mapping.event.ReactiveBeforeConvertCallback;
 import org.springframework.data.mongodb.core.mapping.event.ReactiveBeforeSaveCallback;
-import org.springframework.data.mongodb.repository.support.QuerydslMongoPredicateExecutor;
-import org.springframework.data.mongodb.repository.support.ReactiveQuerydslMongoPredicateExecutor;
 import org.springframework.lang.Nullable;
 import org.springframework.util.ClassUtils;
 
@@ -66,7 +64,6 @@ class MongoRuntimeHints implements RuntimeHintsRegistrar {
 									MemberCategory.INVOKE_PUBLIC_METHODS));
 		}
 
-		registerQuerydslHints(hints, classLoader);
 	}
 
 	private static void registerTransactionProxyHints(RuntimeHints hints, @Nullable ClassLoader classLoader) {
@@ -83,27 +80,4 @@ class MongoRuntimeHints implements RuntimeHintsRegistrar {
 		}
 	}
 
-	/**
-	 * Register hints for Querydsl integration.
-	 *
-	 * @param hints must not be {@literal null}.
-	 * @param classLoader can be {@literal null}.
-	 * @since 4.0.1
-	 */
-	private static void registerQuerydslHints(RuntimeHints hints, @Nullable ClassLoader classLoader) {
-
-		if (ClassUtils.isPresent("com.querydsl.core.types.Predicate", classLoader)) {
-
-			if (isReactorPresent()) {
-				hints.reflection().registerType(ReactiveQuerydslMongoPredicateExecutor.class,
-						MemberCategory.INVOKE_PUBLIC_METHODS, MemberCategory.INVOKE_DECLARED_CONSTRUCTORS);
-
-			}
-
-			if (MongoAotPredicates.isSyncClientPresent(classLoader)) {
-				hints.reflection().registerType(QuerydslMongoPredicateExecutor.class, MemberCategory.INVOKE_PUBLIC_METHODS,
-						MemberCategory.INVOKE_DECLARED_CONSTRUCTORS);
-			}
-		}
-	}
 }
