+++ spring-data-mongodb/src/main/java/org/springframework/data/mongodb/repository/aot/RepositoryRuntimeHints.java
@@ -17,13 +17,15 @@ package org.springframework.data.mongodb.repository.aot;
 
 import static org.springframework.data.mongodb.aot.MongoAotPredicates.*;
 
-import java.util.Arrays;
+import java.util.List;
 
 import org.springframework.aot.hint.MemberCategory;
 import org.springframework.aot.hint.RuntimeHints;
 import org.springframework.aot.hint.RuntimeHintsRegistrar;
 import org.springframework.aot.hint.TypeReference;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
+import org.springframework.data.mongodb.aot.MongoAotPredicates;
+import org.springframework.data.mongodb.repository.support.QuerydslMongoPredicateExecutor;
+import org.springframework.data.mongodb.repository.support.ReactiveQuerydslMongoPredicateExecutor;
 import org.springframework.data.querydsl.QuerydslUtils;
 import org.springframework.lang.Nullable;
 
@@ -37,25 +39,42 @@ class RepositoryRuntimeHints implements RuntimeHintsRegistrar {
 	public void registerHints(RuntimeHints hints, @Nullable ClassLoader classLoader) {
 
 		hints.reflection().registerTypes(
-				Arrays.asList(TypeReference.of("org.springframework.data.mongodb.repository.support.SimpleMongoRepository")),
+				List.of(TypeReference.of("org.springframework.data.mongodb.repository.support.SimpleMongoRepository")),
 				builder -> builder.withMembers(MemberCategory.INVOKE_DECLARED_CONSTRUCTORS,
 						MemberCategory.INVOKE_PUBLIC_METHODS));
 
 		if (isReactorPresent()) {
 
 			hints.reflection().registerTypes(
-					Arrays.asList(
+					List.of(
 							TypeReference.of("org.springframework.data.mongodb.repository.support.SimpleReactiveMongoRepository")),
 					builder -> builder.withMembers(MemberCategory.INVOKE_DECLARED_CONSTRUCTORS,
 							MemberCategory.INVOKE_PUBLIC_METHODS));
 		}
 
 		if (QuerydslUtils.QUERY_DSL_PRESENT) {
+			registerQuerydslHints(hints, classLoader);
+		}
+	}
 
-			hints.reflection().registerType(
-					TypeReference.of("org.springframework.data.mongodb.repository.support.QuerydslMongoPredicateExecutor"),
-					hint -> hint.withMembers(MemberCategory.INVOKE_PUBLIC_CONSTRUCTORS, MemberCategory.INVOKE_PUBLIC_METHODS)
-							.onReachableType(QuerydslPredicateExecutor.class));
+	/**
+	 * Register hints for Querydsl integration.
+	 *
+	 * @param hints must not be {@literal null}.
+	 * @param classLoader can be {@literal null}.
+	 * @since 4.0.2
+	 */
+	private static void registerQuerydslHints(RuntimeHints hints, @Nullable ClassLoader classLoader) {
+
+		if (isReactorPresent()) {
+			hints.reflection().registerType(ReactiveQuerydslMongoPredicateExecutor.class,
+					MemberCategory.INVOKE_PUBLIC_METHODS, MemberCategory.INVOKE_DECLARED_CONSTRUCTORS);
+
+		}
+
+		if (MongoAotPredicates.isSyncClientPresent(classLoader)) {
+			hints.reflection().registerType(QuerydslMongoPredicateExecutor.class, MemberCategory.INVOKE_PUBLIC_METHODS,
+					MemberCategory.INVOKE_DECLARED_CONSTRUCTORS);
 		}
 	}
 }
